+++ src/main/kotlin/dev/aaronhowser/mods/geneticsresequenced/event/custom/ModifyEntityGenesEvent.kt
@@ -1,0 +1,11 @@
+package dev.aaronhowser.mods.geneticsresequenced.event.custom
+
+import dev.aaronhowser.mods.geneticsresequenced.gene.Gene
+import net.minecraft.resources.ResourceKey
+import net.minecraft.world.entity.EntityType
+import net.neoforged.bus.api.Event
+
+data class ModifyEntityGenesEvent(
+	val entityType: ResourceKey<EntityType<*>>,
+	val geneWeights: MutableMap<ResourceKey<Gene>, Int>
+) : Event()