+++ src/main/kotlin/dev/aaronhowser/mods/geneticsresequenced/attachment/GenesData.kt
@@ -1,7 +1,7 @@
 package dev.aaronhowser.mods.geneticsresequenced.attachment
 
 import com.mojang.serialization.Codec
-import dev.aaronhowser.mods.aaron.AaronExtensions.isHolder
+import dev.aaronhowser.mods.aaron.misc.AaronExtensions.isHolder
 import dev.aaronhowser.mods.geneticsresequenced.GeneticsResequenced
 import dev.aaronhowser.mods.geneticsresequenced.attachment.TemporaryGenesData.Companion.temporaryGeneHolders
 import dev.aaronhowser.mods.geneticsresequenced.config.ServerConfig