+++ src/main/kotlin/dev/aaronhowser/mods/geneticsresequenced/block/block_entity/DnaDecryptorBlockEntity.kt
@@ -1,7 +1,7 @@
 package dev.aaronhowser.mods.geneticsresequenced.block.block_entity
 
-import dev.aaronhowser.mods.aaron.AaronExtensions.isItem
-import dev.aaronhowser.mods.aaron.ImprovedSimpleContainer
+import dev.aaronhowser.mods.aaron.misc.AaronExtensions.isItem
+import dev.aaronhowser.mods.aaron.misc.ImprovedSimpleContainer
 import dev.aaronhowser.mods.geneticsresequenced.block.base.CraftingMachineBlockEntity
 import dev.aaronhowser.mods.geneticsresequenced.data.EntityGenes
 import dev.aaronhowser.mods.geneticsresequenced.gene.Gene