+++ src/main/kotlin/dev/aaronhowser/mods/geneticsresequenced/gene/Gene.kt
@@ -2,11 +2,11 @@ package dev.aaronhowser.mods.geneticsresequenced.gene
 
 import com.mojang.serialization.Codec
 import com.mojang.serialization.codecs.RecordCodecBuilder
-import dev.aaronhowser.mods.aaron.AaronExtensions.isHolder
-import dev.aaronhowser.mods.aaron.AaronExtensions.tell
-import dev.aaronhowser.mods.aaron.AaronExtensions.withClickToCopyToClipboard
-import dev.aaronhowser.mods.aaron.AaronExtensions.withHoverText
-import dev.aaronhowser.mods.aaron.AaronExtraCodecs
+import dev.aaronhowser.mods.aaron.misc.AaronExtensions.isHolder
+import dev.aaronhowser.mods.aaron.misc.AaronExtensions.tell
+import dev.aaronhowser.mods.aaron.misc.AaronExtensions.withClickToCopyToClipboard
+import dev.aaronhowser.mods.aaron.misc.AaronExtensions.withHoverText
+import dev.aaronhowser.mods.aaron.serialization.AaronExtraCodecs
 import dev.aaronhowser.mods.geneticsresequenced.GeneticsResequenced
 import dev.aaronhowser.mods.geneticsresequenced.datagen.gene.ModGeneProvider
 import dev.aaronhowser.mods.geneticsresequenced.datagen.lang.ModGeneLang