+++ src/main/kotlin/dev/aaronhowser/mods/geneticsresequenced/menu/CraftingMachineMenu.kt
@@ -1,8 +1,8 @@
 package dev.aaronhowser.mods.geneticsresequenced.menu
 
-import dev.aaronhowser.mods.aaron.AaronExtensions.isItem
 import dev.aaronhowser.mods.aaron.menu.components.FilteredSlot
 import dev.aaronhowser.mods.aaron.menu.components.OutputSlot
+import dev.aaronhowser.mods.aaron.misc.AaronExtensions.isItem
 import dev.aaronhowser.mods.geneticsresequenced.block.base.CraftingMachineBlockEntity
 import dev.aaronhowser.mods.geneticsresequenced.block.base.container_data.CraftingContainerData
 import dev.aaronhowser.mods.geneticsresequenced.registry.ModItems