+++ src/main/kotlin/dev/aaronhowser/mods/geneticsresequenced/recipe/crafting/SetAntiPlasmidRecipe.kt
@@ -1,6 +1,6 @@
 package dev.aaronhowser.mods.geneticsresequenced.recipe.crafting
 
-import dev.aaronhowser.mods.aaron.AaronExtensions.withComponent
+import dev.aaronhowser.mods.aaron.misc.AaronExtensions.withComponent
 import dev.aaronhowser.mods.geneticsresequenced.item.PlasmidItem
 import dev.aaronhowser.mods.geneticsresequenced.registry.ModDataComponents
 import dev.aaronhowser.mods.geneticsresequenced.registry.ModItems