+++ CHANGELOG.md
@@ -10,6 +10,19 @@
 	- A regular Cell will craft 8
 	- A GMO Cell will craft 4
 - Gene Requirements are now a datapack registry
+	- The json is still in `/data/_____/geneticsresequenced/gene_requirements/`, but the file name has to be the gene id
+	- That is, `geneticsresequenced:flight` has to go in `/data/geneticsresequenced/geneticsresequenced/gene_requirements/flight.json`
+	- Inside the json is a list of required Gene IDs, like so:
+
+```json
+{
+	"requirements": [
+		"geneticsresequenced:teleport",
+		"geneticsresequenced:step_assist",
+		"geneticsresequenced:no_fall_damage"
+	]
+}
+```
 
 ### Fixed